A snake game working on console window with many game levels and options
Up to now this game is working on the console window. several game levels were developed. there are different options as well. all functionality are handled by game loop. Inside the loop we have several game functions. Functions were developed in outside the main function and functions are called in game loop that is in main function.